Service layer, Exception handler P.S. Repository names without "local" :P

+++ backend/rest/MerchantRest/src/main/java/app/services/RealestateService.java
@@ -1,37 +1,79 @@
 package app.services;
 
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import app.model.RealestateInsPackage;
-import app.repository.local.RealestatePackLocalRepository;
+import app.repository.local.RealestatePackRepository;
+import app.services.exceptions.BadRequestException;
+import app.services.exceptions.NotFoundException;
 
 @Service
 public class RealestateService {
+	
 	@Autowired
-	private RealestatePackLocalRepository realestateRepository;
+	private RealestatePackRepository realestateRepository;
+	
+	private static final Logger logger = Logger.getLogger(RealestateService.class);
 
 	
-	public RealestateInsPackage create(RealestateInsPackage realestate) {
-		return realestateRepository.save(realestate);
+	public Map<String, Object> create(RealestateInsPackage realestate) {
+		
+		Map<String, Object> response;
+		
+		if (realestate == null) {
+			throw new BadRequestException();		
+		}
+		
+		response = new LinkedHashMap<String, Object>();
+		realestateRepository.save(realestate);
+		response.put("message", "Package created successfully");
+		response.put("package", realestate);
+		logger.info("Package created successfully");
+		
+		return response;
 	
 	}
 
 
 	public List<RealestateInsPackage> findAll() {
+		logger.info("List all packages");
 		return (List<RealestateInsPackage>) realestateRepository.findAll();
 	}
 
 	
 	public RealestateInsPackage findById( String packageId) {
-		return realestateRepository.findOne(packageId);
+		
+		if(packageId == null)
+		{
+			throw new BadRequestException();
+		} 
+		
+		logger.info("Find package with id: " + packageId);
+		RealestateInsPackage rPackage =realestateRepository.findOne(packageId);
+		
+		if (rPackage == null) {
+			throw new NotFoundException();
+		}
+		return rPackage;
 	}
 
 	
-	public void remove(String packageId) {
-		realestateRepository.delete(packageId);
+	public String remove(String packageId) {
+
+		if(packageId == null)
+		{
+			throw new BadRequestException();
+		}
+		logger.info("Removing package with id: " + packageId);		
+		realestateRepository.delete(packageId);	
+		return "removed";
+	
 
 	}
 }
